Properly seperate modules and merge Dockerfile
