+++ .editorconfig
@@ -21,3 +21,4 @@ indent_size = 4
 # Java
 [*.java]
 indent_size = 4
+continuation_indent_size = 8
