+++ java-sdk/radar-schemas-commons/build.gradle
@@ -1,10 +1,28 @@
-ext.artifactName = 'radar-schemas-commons'
+plugins {
+    id 'java-library'
+    id 'com.github.davidmc24.gradle.plugin.avro-base'
+}
+
 description = 'RADAR Schemas Commons SDK'
 
 sourceSets {
     main {
         java.srcDir avroOutputDir
     }
+}
+
+dependencies {
+    api (group: 'org.apache.avro', name: 'avro', version: avroVersion) {
+        exclude group: 'org.xerial.snappy', module: 'snappy-java'
+        exclude group: 'com.thoughtworks.paranamer', module: 'paranamer'
+        exclude group: 'org.apache.commons', module: 'commons-compress'
+        exclude group: 'org.tukaani', module: 'xz'
+    }
+}
+
+compileJava {
+    targetCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = JavaVersion.VERSION_1_8
 }
 
 //---------------------------------------------------------------------------//
@@ -17,7 +35,7 @@ clean {
 //---------------------------------------------------------------------------//
 // AVRO file manipulation                                                    //
 //---------------------------------------------------------------------------//
-task generateAvro(type: com.commercehub.gradle.plugin.avro.GenerateAvroJavaTask) {
+task generateAvro(type: com.github.davidmc24.gradle.plugin.avro.GenerateAvroJavaTask) {
     source rootProject.fileTree('../commons') {
         include '**/*.avsc'
     }
