+++ java-sdk/radar-schemas-core/src/main/java/org/radarbase/schema/specification/AppDataTopic.java
@@ -2,10 +2,13 @@ package org.radarbase.schema.specification;
 
 import static org.radarbase.schema.util.SchemaUtils.expandClass;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSetter;
 import java.util.Map;
 
+@JsonInclude(Include.NON_NULL)
 public class AppDataTopic extends DataTopic {
     @JsonProperty("app_provider")
     private String appProvider;
