+++ java-sdk/radar-schemas-core/src/main/java/org/radarbase/schema/specification/AppSource.java
@@ -2,10 +2,13 @@ package org.radarbase.schema.specification;
 
 import static org.radarbase.schema.util.SchemaUtils.expandClass;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSetter;
 import java.util.Objects;
 
+@JsonInclude(Include.NON_NULL)
 public abstract class AppSource<T extends DataTopic> extends DataProducer<T> {
     @JsonProperty("app_provider")
     private String appProvider;
