+++ java-sdk/radar-schemas-core/src/main/java/org/radarbase/schema/specification/monitor/MonitorSource.java
@@ -1,11 +1,14 @@
 package org.radarbase.schema.specification.monitor;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 import org.radarbase.schema.Scope;
 import org.radarbase.schema.specification.AppDataTopic;
 import org.radarbase.schema.specification.AppSource;
 
+@JsonInclude(Include.NON_NULL)
 public class MonitorSource extends AppSource<AppDataTopic> {
     @JsonProperty
     private List<AppDataTopic> data;
