+++ java-sdk/radar-schemas-registration/README.md
@@ -7,6 +7,7 @@ the tools can perform.
 The RADAR Schemas Validator checks if the `Schema Catalog` is in a valid state.
 
 It first checks the folder structure, it has to be compliant with:
+
 - commons
   * active
   * catalogue
@@ -20,8 +21,9 @@ It first checks the folder structure, it has to be compliant with:
   * monitor
   * passive
   * stream
-  
+
 For each Avro schema under `commons` folder checks if:
+
 - in case of `ENUM`
   * the `namespace` differs from null and it is a lowercase string dot separated without numeric
   * the `name` matches the .avsc file name and it is an UpperCamelCase string
@@ -31,14 +33,18 @@ For each Avro schema under `commons` folder checks if:
   * the `namespace` differs from null and it is a lowercase string dot separated without numeric
   * the `name` matches the .avsc file name and it is an UpperCamelCase string
   * fields is not empty
-  * the `field name` is a lowerCamelCase string and does not contain string such as `value`, `Value`.
-  * schemas under `active` folder have `time` and `timeCompleted` fields, and do not contain a field named `timeReceived`
-  * schemas under `monitor` folder have `time` field, and do not contain a field named either `timeCompleted` or `timeReceived`
-  * schemas under `passive` folder have `time` and `timeReceived` fields, and do not have a field named `timeCompleted`
+  * the `field name` is a lowerCamelCase string and does not contain string such as `value`, `Value`
+    .
+  * schemas under `active` folder have `time` and `timeCompleted` fields, and do not contain a field
+    named `timeReceived`
+  * schemas under `monitor` folder have `time` field, and do not contain a field named
+    either `timeCompleted` or `timeReceived`
+  * schemas under `passive` folder have `time` and `timeReceived` fields, and do not have a field
+    named `timeCompleted`
   * the record and any provided fields are documented
   * `ENUM` fields have `UNKNOWN` as `default` value, if any
-  * `nullable`/`optional` fields have `null` as default value 
-  
+  * `nullable`/`optional` fields have `null` as default value
+
 Upon rule violation, the end user is notified with a message explaining how to fix it.
 
 ## How to use
@@ -47,12 +53,16 @@ The validation is implemented as a `JUnit` test. To run the validation, simply t
 
 ## Suppress checks
 
-Record name and field name validations can be suppressed modifying the [skip](src/test/resources/schema.yml) configuration file.
+Record name and field name validations can be suppressed modifying
+the [skip](src/test/resources/schema.yml) configuration file.
 
 `files` lists file paths that can be ignored. It can contain values like
+
 - entire path like `commons/active/questionnaire/questionnaire.avsc`
-- folder and subfolder `commons/active/**/*`: all file under `active` and all its subfolder will be skipped
-- folder and subfolder `commons/active/**/*.avsc`: all file with format `avsc` under `active` and all its subfolder will be skipped
+- folder and subfolder `commons/active/**/*`: all file under `active` and all its subfolder will be
+  skipped
+- folder and subfolder `commons/active/**/*.avsc`: all file with format `avsc` under `active` and
+  all its subfolder will be skipped
 - file name `.DS_Store`: all file named `.DS_Store` will be skipped
 - file extension `*.md`: all file with extension `*.md` will be skipped
 
@@ -66,11 +76,14 @@ files:
 ``` 
 
 `validation` allows the user to specify checks that should be skipped at schema level:
-- a key like `org.radarcns.passive.biovotion.*` set a skip configuration valid for all schemas under `org.radarcns.passive.biovotion` package
-- a key like `org.radarcns.passive.biovotion.BiovotionVSMSpO2` specify a configuration valid only for the given record
+
+- a key like `org.radarcns.passive.biovotion.*` set a skip configuration valid for all schemas
+  under `org.radarcns.passive.biovotion` package
+- a key like `org.radarcns.passive.biovotion.BiovotionVSMSpO2` specify a configuration valid only
+  for the given record
 - `name_record_check: DISABLE` suppresses the record name check
-- `fields` lists field names for which the field name check is suppressed
-All tests are enable by default.
+- `fields` lists field names for which the field name check is suppressed All tests are enable by
+  default.
 
 ```yaml
 validation:
