+++ java-sdk/radar-schemas-registration/src/main/java/org/radarbase/schema/registration/SchemaRegistry.java
@@ -16,19 +16,10 @@
 
 package org.radarbase.schema.registration;
 
-import static org.radarbase.util.Strings.isNullOrEmpty;
-import static org.radarbase.schema.CommandLineApp.matchTopic;
-
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.util.Comparator;
-import java.util.Optional;
 import java.util.concurrent.TimeUnit;
-import java.util.regex.Pattern;
-
-import net.sourceforge.argparse4j.impl.Arguments;
-import net.sourceforge.argparse4j.inf.ArgumentParser;
-import net.sourceforge.argparse4j.inf.Namespace;
 import okhttp3.Credentials;
 import okhttp3.Headers;
 import okhttp3.MediaType;
@@ -41,11 +32,9 @@ import org.jetbrains.annotations.NotNull;
 import org.radarbase.config.ServerConfig;
 import org.radarbase.producer.rest.RestClient;
 import org.radarbase.producer.rest.SchemaRetriever;
-import org.radarbase.topic.AvroTopic;
-import org.radarbase.schema.CommandLineApp;
 import org.radarbase.schema.specification.DataProducer;
 import org.radarbase.schema.specification.SourceCatalogue;
-import org.radarbase.schema.util.SubCommand;
+import org.radarbase.topic.AvroTopic;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -53,16 +42,13 @@ import org.slf4j.LoggerFactory;
  * Schema registry interface.
  */
 public class SchemaRegistry {
-    public enum Compatibility {
-        NONE, FULL, BACKWARD, FORWARD, BACKWARD_TRANSITIVE, FORWARD_TRANSITIVE, FULL_TRANSITIVE
-    }
-
     private static final Logger logger = LoggerFactory.getLogger(SchemaRegistry.class);
     private final SchemaRetriever schemaClient;
     private final RestClient httpClient;
 
     /**
      * Schema registry for given URL. If this is https, unsafe certificates are accepted.
+     *
      * @param baseUrl URL of the schema registry
      * @throws MalformedURLException if given URL is invalid.
      */
@@ -70,9 +56,9 @@ public class SchemaRegistry {
         ServerConfig config = new ServerConfig(baseUrl);
         config.setUnsafe(false);
         this.httpClient = RestClient.global()
-            .timeout(10, TimeUnit.SECONDS)
-            .server(config)
-            .build();
+                .timeout(10, TimeUnit.SECONDS)
+                .server(config)
+                .build();
         this.schemaClient = new SchemaRetriever(this.httpClient);
     }
 
@@ -90,6 +76,7 @@ public class SchemaRegistry {
 
     /**
      * Register all schemas in a source catalogue. Stream and connector sources are ignored.
+     *
      * @param catalogue schema catalogue to read schemas from
      * @return whether all schemas were successfully registered.
      */
@@ -105,7 +92,9 @@ public class SchemaRegistry {
                 .allMatch(this::registerSchema);
     }
 
-    /** Register the schema of a single topic. */
+    /**
+     * Register the schema of a single topic.
+     */
     public boolean registerSchema(AvroTopic<?, ?> topic) {
         try {
             this.schemaClient.addSchema(topic.getName(), false, topic.getKeySchema());
@@ -119,6 +108,7 @@ public class SchemaRegistry {
 
     /**
      * Set the compatibility level of the schema registry.
+     *
      * @param compatibility target compatibility level.
      * @return whether the request was successful.
      */
@@ -164,88 +154,7 @@ public class SchemaRegistry {
         }
     }
 
-    /** Return the schema registry as a subcommand. */
-    public static SubCommand command() {
-        return new RegisterCommand();
-    }
-
-    private static class RegisterCommand implements SubCommand {
-        @Override
-        public String getName() {
-            return "register";
-        }
-
-        @Override
-        public int execute(Namespace options, CommandLineApp app) {
-            String url = options.get("schemaRegistry");
-            String apiKey = options.getString("api_key");
-            String apiSecret = options.getString("api_secret");
-            try {
-                SchemaRegistry registration;
-                if (isNullOrEmpty(apiKey) || isNullOrEmpty(apiSecret)) {
-                    logger.info("Initializing standard SchemaRegistration ...");
-                    registration = new SchemaRegistry(url);
-                } else {
-                    logger.info("Initializing SchemaRegistration with authentication...");
-                    registration = new SchemaRegistry(url, apiKey, apiSecret);
-                }
-
-                boolean forced = options.getBoolean("force");
-                if (forced && !registration.putCompatibility(Compatibility.NONE)) {
-                    return 1;
-                }
-                boolean result;
-                Pattern pattern = matchTopic(
-                        options.getString("topic"), options.getString("match"));
-
-                if (pattern == null) {
-                    result = registration.registerSchemas(app.getCatalogue());
-                } else {
-                    Optional<Boolean> didUpload = app.getCatalogue().getTopics()
-                            .filter(t -> pattern.matcher(t.getName()).find())
-                            .map(registration::registerSchema)
-                            .reduce((a, b) -> a && b);
-
-                    if (didUpload.isPresent()) {
-                        result = didUpload.get();
-                    } else {
-                        logger.error("Topic {} does not match a known topic."
-                                + " Find the list of acceptable topics"
-                                + " with the `radar-schemas-tools list` command. Aborting.",
-                                pattern);
-                        result = false;
-                    }
-                }
-                if (forced) {
-                    registration.putCompatibility(Compatibility.FULL);
-                }
-                return result ? 0 : 1;
-            } catch (MalformedURLException ex) {
-                logger.error("Schema registry URL {} is invalid: {}", url, ex.toString());
-                return 1;
-            }
-        }
-
-        @Override
-        public void addParser(ArgumentParser parser) {
-            parser.description("Register schemas in the schema registry.");
-            parser.addArgument("-f", "--force")
-                    .help("force registering schema, even if it is incompatible")
-                    .action(Arguments.storeTrue());
-            parser.addArgument("-t", "--topic")
-                    .help("register the schemas of one topic")
-                    .type(String.class);
-            parser.addArgument("-m", "--match")
-                    .help("register the schemas of all topics matching the given regex"
-                            + "; does not do anything if --topic is specified")
-                    .type(String.class);
-            parser.addArgument("schemaRegistry")
-                    .help("schema registry URL");
-            parser.addArgument("-u", "--api-key")
-                    .help("Client password to authorize with.");
-            parser.addArgument("-p", "--api-secret")
-                    .help("Client key to authorize with.");
-            SubCommand.addRootArgument(parser);
-        }
+    public enum Compatibility {
+        NONE, FULL, BACKWARD, FORWARD, BACKWARD_TRANSITIVE, FORWARD_TRANSITIVE, FULL_TRANSITIVE
     }
 }
