+++ java-sdk/radar-schemas-registration/src/test/resources/key_measurement_test.avsc
