+++ java-sdk/radar-schemas-registration/src/test/resources/key_windowed_test.avsc
