+++ java-sdk/radar-schemas-registration/src/test/resources/logback.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+
+  <appender class="ch.qos.logback.core.ConsoleAppender" name="STDOUT">
+    <encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
+      <pattern>%date{yyyy-MM-dd HH:mm:ss,UTC} UTC %-5level - %msg [%logger:%line] %n</pattern>
+    </encoder>
+  </appender>
+
+  <logger additivity="false" level="info" name="org.radarbase.validator">
+    <appender-ref ref="STDOUT"/>
+  </logger>
+
+  <root level="info">
+    <appender-ref ref="STDOUT"/>
+  </root>
+</configuration>
