+++ java-sdk/radar-schemas-registration/src/test/resources/schema.yml
