--- java-sdk/radar-schemas-tools/src/test/resources/logback.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-
-    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder class="ch.qos.logback.classic.encoder.PatternLayoutEncoder">
-            <pattern>%date{yyyy-MM-dd HH:mm:ss,UTC} UTC %-5level - %msg [%logger:%line] %n</pattern>
-        </encoder>
-    </appender>
-
-    <logger name="org.radarcns.validator" level="info" additivity="false">
-        <appender-ref ref="STDOUT" />
-    </logger>
-
-    <root level="info">
-        <appender-ref ref="STDOUT" />
-    </root>
-</configuration>
